requirements.txt
duckdb
flask
python-dotenv
